refactor: move proyectos components and utils
