+++ src/app/ui/button/index.ts
@@ -1,4 +1,3 @@
-export * from "./CreateProyectoButton"
 export * from "./DeleteButton"
 export * from "./EditButton"
 export * from "./SearchProyectosByInvestigadorButton"
