+++ src/app/ui/cards/index.ts
@@ -1,4 +1,3 @@
 export * from "./CardModal"
 export * from "./CardSkeleton"
 export * from "./HorizontalCard"
-export * from "./proyecto"
