--- src/app/ui/cards/proyecto/SubmitCard.tsx
@@ -1,13 +1,0 @@
-import { ProyectoFileUploaderForm } from "../../form/ProyectoFileUploadForm"
-import { CardModal } from "../CardModal"
-
-export const ProyectoSubmitCard = ({ onClose }: { onClose: () => void }) => {
-  return (
-    <CardModal onClose={onClose}>
-      <h2 className="text-2xl text-accent-primary">
-        Añadir Proyectos por archivo
-      </h2>
-      <ProyectoFileUploaderForm />
-    </CardModal>
-  )
-}
