+++ src/app/ui/form/index.ts
@@ -1,4 +1,2 @@
-export * from "./create"
-export * from "./edit"
 // eslint-disable-next-line import/no-cycle
 export * from "./ProyectoFileUploadForm"
