+++ src/app/ui/proyectoByInvestigador/MostrarProyectos.tsx
@@ -9,10 +9,13 @@ import { useEffect, useRef, useState } from "react"
 
 import { useQueryParam } from "@/app/utils/hooks/useQueryParam"
 import { fetchProyectoByCode } from "@/db"
-import { ProyectoCard, ProyectoMiniCard } from "@/proyectos"
+import type { ProyectoMinimumDataType, ProyectoType } from "@/proyectos"
+import {
+  downloadProyectosCSV,
+  ProyectoCard,
+  ProyectoMiniCard,
+} from "@/proyectos"
 
-import type { ProyectoMinimumDataType, ProyectoType } from "../../utils"
-import { downloadProyectosCSV } from "../../utils"
 import { Button } from "../button/Button"
 import { InvestigadoresList } from "./InvestigadoresList"
 import {
