+++ src/app/utils/types/items.ts
@@ -1,15 +1,3 @@
-export type ProyectoType = {
-  codigo: string
-  ip: string
-  coip?: string
-  titulo: string
-  financiado: string
-  inicio: Date
-  fin?: Date
-}
-
-export type ProyectoMinimumDataType = Pick<ProyectoType, "codigo" | "titulo">
-
 export type ParticipaType = {
   codigo: string
   email: string
