--- src/app/utils/validation/index.ts
@@ -1,2 +1,0 @@
-export * from "./constants"
-export * from "./project"
