+++ src/proyectos/cards/ProyectoUploadCard.tsx
@@ -10,23 +10,21 @@ import cx from "classnames"
 import type { KeyboardEvent } from "react"
 import { useState } from "react"
 
-import type {
-  ErrorMessage,
-  ParticipaType,
-  ProyectoToUpload,
-  ProyectoType,
-} from "@/app/utils"
+import type { ParticipaType } from "@/app/utils"
+
+import { EditButton } from "../../app/ui/button"
+import { HorizontalCard } from "../../app/ui/cards/HorizontalCard"
+// eslint-disable-next-line import/no-cycle
+import { ErrorMessageCard } from "../../app/ui/information/proyectoFileUpdate/messages/ErrorMessageCard"
+import { WarningMessageCard } from "../../app/ui/information/proyectoFileUpdate/messages/WarningMessageCard"
+import { EditProyectoForm } from "../form"
+import type { ProyectoType } from "../types"
+import type { ErrorMessage } from "../utils/infoMessage"
+import type { ProyectoToUpload } from "../utils/map"
 import {
   mapProyectoToUploadToProyectType,
   mapProyectoTypeToProyectoToUpload,
-} from "@/app/utils"
-
-import { EditButton } from "../../button"
-// eslint-disable-next-line import/no-cycle
-import { EditProyectoForm } from "../../form"
-import { ErrorMessageCard } from "../../information/proyectoFileUpdate/messages/ErrorMessageCard"
-import { WarningMessageCard } from "../../information/proyectoFileUpdate/messages/WarningMessageCard"
-import { HorizontalCard } from "../HorizontalCard"
+} from "../utils/map"
 
 interface ProyectoCardToUploadProps {
   proyecto: ProyectoToUpload
