+++ src/proyectos/cards/index.ts
@@ -1,2 +1,3 @@
 export * from "./ProyectoCard"
 export * from "./ProyectoMiniCard"
+export * from "./ProyectoUploadCard"
