+++ src/proyectos/form/create/ProyectoCreateForm.tsx
@@ -7,9 +7,9 @@ import { useState } from "react"
 
 import { parseDateToString } from "@/app/utils"
 
-import { Button } from "../../button/Button"
-import { HorizontalCard } from "../../cards/HorizontalCard"
-import { AddParticipanteCard } from "../../menu/AddParticipante"
+import { Button } from "../../../app/ui/button/Button"
+import { HorizontalCard } from "../../../app/ui/cards/HorizontalCard"
+import { AddParticipanteCard } from "../../../app/ui/menu/AddParticipante"
 import { useProyectoCreate } from "./useProyectoCreate"
 
 export const ProyectoCreate = () => {
