+++ src/proyectos/index.ts
@@ -1,0 +1,5 @@
+export * from "./button"
+export * from "./cards"
+export * from "./form"
+export * from "./types"
+export * from "./utils"
