+++ src/proyectos/utils/addAll.ts
