+++ src/proyectos/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./addAll"
 export * from "./fetch"
 export * from "./infoMessage"
 export * from "./map"
+export * from "./validation"
